third_party/blink/web_tests/fast/forms/state-restore-broken-state-expected.txt

CONSOLE MESSAGE: Generated state: [state-restore-broken-state-2.html [name1 ] #0,1,name1,text,1,modified]
The value was modified in the first load of state-restore-broken-state-1.html, but it should not be restored because the state-restore-broken-state-2.html breaks the state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementsByTagName("iframe")[0].contentDocument.f.name1.value is "unmodified"
PASS successfullyParsed is true

TEST COMPLETE
